share common query service code
